backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Vector DB
qdrant-client==1.7.1

# AI/ML
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.8.1
langchain==0.1.6
langchain-community==0.0.20
langchain-openai==0.0.5

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
pyyaml==6.0.1
markdown==3.5.2
beautifulsoup4==4.12.3
asyncpg==0.29.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# CORS
fastapi-cors==0.0.6
